feat: enhance scrobbling modal
